Separate test_legacy_machinetype as test_legacy_machinetype.py
The rhel-8-10 matrix_os impact on ARM64 related tests within machintype cases.
This change ensures that test_legacy_machinetype case will not impact ARM64 test.
